replace data step with refactored one
