+++ src/gz21_ocean_momentum/step/data/coarsen.py
