+++ src/gz21_ocean_momentum/step/data/lib.py
@@ -1,5 +1,4 @@
-import gz21_ocean_momentum.new.data.coarsen as coarsen
-import gz21_ocean_momentum.new.data.utils   as utils
+import gz21_ocean_momentum.step.data.coarsen as coarsen
 
 import xarray as xr
 import intake
@@ -27,9 +26,9 @@ def preprocess_and_compute_forcings(
     if cyclize:
         # TODO logger
         #logger.info("Cyclic data... Making the dataset cyclic along longitude...")
-        surface_fields = utils.cyclize(
+        surface_fields = _cyclize(
                 surface_fields, "xu_ocean", resolution_degrading_factor)
-        grid = utils.cyclize(
+        grid = _cyclize(
                 grid,           "xu_ocean", resolution_degrading_factor)
 
         # rechunk along the cyclized dimension
@@ -58,3 +57,39 @@ def download_cm2_6(
         surface_fields = catalog.ocean.GFDL_CM2_6.GFDL_CM2_6_one_percent_ocean_surface
     surface_fields = surface_fields.to_dask()
     return surface_fields, grid
+
+def _cyclize(ds: xr.Dataset, coord_name: str, nb_points: int):
+    """
+    Generate a cyclic dataset from non-cyclic input.
+
+    Return a cyclic dataset, with nb_points added on each end, along
+    the coordinate specified by coord_name.
+
+    Parameters
+    ----------
+    ds : xr.Dataset
+        Dataset to process.
+    coord_name : str
+        Name of the coordinate along which the data is made cyclic.
+    nb_points : int
+        Number of points added on each end.
+
+    Returns
+    -------
+    New extended dataset.
+    """
+    # TODO make this flexible
+    cycle_length = 360.0
+    left = ds.roll({coord_name: nb_points}, roll_coords=True)
+    right = ds.roll({coord_name: nb_points}, roll_coords=True)
+    right = right.isel({coord_name: slice(0, 2 * nb_points)})
+    left[coord_name] = xr.concat(
+        (left[coord_name][:nb_points] - cycle_length, left[coord_name][nb_points:]),
+        coord_name,
+    )
+    right[coord_name] = xr.concat(
+        (right[coord_name][:nb_points], right[coord_name][nb_points:] + cycle_length),
+        coord_name,
+    )
+    new_ds = xr.concat((left, right), coord_name)
+    return new_ds
